IMPACTS: a big update to get something in there asap, still partly deactive.
This introduces a pre-scene module which holds a texture used for data transfer from CPU to the GPU. It is per-scene and not per-object to minimize Lock/Unlock. This module is passed to all objects in the scene via UpdateContext.

Binaries submitted in 972008
Build time : 9 minutes 16.808000 seconds
Branch : //eve/branches/development/MAINLINE/

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/Trinity/": change = 972003]