[eve MAINLINE] Automatic VCXPROJ submission from 1346264.4398
This might create build loops, but it was deemed valuable enough to be able to sync these.
https://teamcity.evetech.net/viewLog.html?buildId=2892991
!nobuild

Source Changes
----------------
[Eve Engine] 1346309 by markw (Mark Wilkins)
Fixing typo in yaml for build
https://jira.ccpgames.com/browse/TQ-130092
----------------
[Eve Engine] 1346268 by markw (Mark Wilkins)
Adding debug version of squish library
https://jira.ccpgames.com/browse/TQ-130092
----------------
[Eve Engine] 1346253 by huni (H??ni Hinrichsen)
TOOLS: GMs can see player's current sentence
https://jira.ccpgames.com/browse/TP-950
TOOLS: GM can add a sentence for a player
https://jira.ccpgames.com/browse/TP-951
- added ability to add a penalty to a character in GM pages
- created a view of present and passed penalties for a character in GM pages
- created new proc that gets all penalties for a character
Review: http://codecollaborator/ui#review:id=15329
EBS 7936
----------------
[Eve Engine] 1346209 by jamieb (Jamie Bannister)
Allow some structures to have shorter deployment ranges to other structures
https://jira.ccpgames.com/browse/TFO-2651
Structure deployment now handles structure-to-structure range checks slightly differently -
Previously: only the minimum deployment range of the already-in-space structure was considered.
New: we find the minimum deployment ranges of the already-in-space structure and the to-be-deployed structure. The LOWEST of these two ranges is now used as the for the range check.
(For non-structure items in space, the range check is unchanged - the structure's minimum deployment range is not used)
Added a (placeholder) minimum range for the StructureNavigation group (1408 - previously known as the StructureStargate group) of 200km. All other structures will continue to use the Structure category's range of 1000km.
Review: http://codecollaborator/ui#review:id=15325
----------------
[Eve Engine] 1346162 by markw (Mark Wilkins)
Updating to support Squish library for x64
https://jira.ccpgames.com/browse/TQ-130092
----------------

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/Trinity/": change = 1346331]