[eve MAINLINE] Automatic VCXPROJ submission from 915
This might create build loops, but it was deemed valuable enough to be able to sync these.
https://teamcity.evetech.net/viewLog.html?buildId=3714823
!nobuild

Source Changes
----------------
[Eve Engine] 1397771 by kristjang (Kristj??n Gerhardsson)
Split CcpCore project into CcpCore and CcpCoreStatic. Some projects depend on the static build while others depend on the non-static build (and some depend on both), so it makes sense to have it as two seperate projects files.
----------------

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/": change = 1397839]