Moving upscaling effects from postprocess to trinityAL
The upscaling technique now lives on the primaryrendercontext and is accessable from the device.
This is done to have better control over the techniques (creating device and swapchains etc).

Each technique creates an upscaling context that is used to render the upscaled image to the screen.

Tr2Streamline has been removed and the logic for interacting with it has been moved to the dlss upscalers.

MetalFx and Fsr1 remain..
https://ccpgames.atlassian.net/browse/PLAT-5326