FSR3 upscaling with framegeneration (doesn't work quite yet)
Added  FSR3 with framegen which crashes still. (need to get some other stuff in right now).

made the MarkFrameEvent func take the rendercontext as a parameter, so we can trigger framegen on FSR3.

Updated the swapchain to be swapchain4 so we can use FSR3 with framegen.

Fixed the renderfps as it had switched temporal and framegeneration