Fix UnicodeDecodeError in `Tr2MainWindow::GetKeyName` (#6)
The Python exposure for this method was failing because the method would return a string constructed based on the Windows code pages, but the exposure expects a utf-8 encoded string. Therefore, use the Unicode versions of Windows API calls and return a std::wstring containing a utf-16 encoded string.

In Python 2.7 the case that is now failing would result in raw bytes being returned, but those were not decoded as utf-8 so no error was raised.