Fixing a few errors with dlss
Removed the GetAdapter function, since it needed a pointer to data that would be deleted. Also since that function was only called once, it can be in the calling position without breaking any DRY rules.

Made the Tr2StreamlineAL::IsFrameGenerationAvailable set the m_supportsFrameGeneration bool, but not the m_frameGeneration bool (supportsFrameGeneration would always be false!)