Fix UnicodeDecodeError in Trinity's GetKeyName
BlueExposure expects a UTF-8 encoded string, but that's not what was getting returned. Making sure we convert the characters we get from UCKeyTranslate to a proper UTF-8 string, which is allowed to contain more than one character.

Also fixing up a string constant that was set to a value that was not valid UTF-8