New bloom (#96)
* Downsample and upsample bloom

New bloom method which downsamples the rendered image and upsamples it again.

There are a few nasty things in here (especially in EveSpaceScene) but I will iron them out after the refactorings to the Volumes is done.

https://ccpgames.atlassian.net/browse/PLAT-8492

* New Bloom - controllable blur

Added blur kernels into the perobject data
Added better tinting
Added better debug mode
Added directional weights

* Improving the CalculateGaussianParameters

Now the CalculateGaussianParameters will take the weight (or tint) and direction as parameters, instead of having to go through the loop a 3 times, we do it at most 2 times.

Also made the new bloom the default

* Adding new bloom to tr2postprocessattributes

All the new values for Bloom are now exposed in Tr2PostProcessAttributes
Added a way to set a default value if there is nothing that modifies the value. This is to make sure that we don't force the user to set the default values every time.

Also fixed a thing where if you set any colorCorrection attribute to something and then disable it, then it will not get reset to the default value.

* Cleaning up

Made the downsampler have perobject data so it is slightly faster
Removed and renamed a few things so they make more sense

* code style fix

* Changing postprocess accumulation logic

I had changed it before but it was a bit dumb, so I' reverting it!

Also removed default values for color correction in the postprocess accumulation, since it is always available on the scenedefaultpostprocess object. (but bloom attributes are not always available...)

* Update Tr2PostProcessAttributes.cpp

* Fixing up postprocess attributes

They are now picking up the new bloom stuff from the postprocess
Also added the resetting of the new bloom stuff to the reset

* Update EveSpaceScene.cpp

* Reverting some postprocess copy function

* reverting some evespacescene newline changes

* Fixing things based review

Changed all bloom defines to consts

Changed the GaussianData to be a packed Vector4 which contains 2 weight and offset pairs.

Moved the useNewBloom and bloom debug to the postprocess step.

Removed dumb code