For multi-hull space objects, decal buffers need to be regenerated on at runtime. (#108)
Added a flag to EveSpaceObjectDecal to force runtime regeneration.
Added an override MapForReading to Tr2Buffer that takes in a size and an offset, so that we don't create huge staging resources every time we need to read a sub-buffer from a suballocated buffer.
Changed how sub-allocated buffer handles read backs: it now reads only the required portion of the buffer, but creates a copy of it in RAM to still support multiple simultanious readbacks.