Deprecate old stretch audio logic and add new implementation (#250)
This commit deprecates the Tr2StretchAudioBase and the Tr2StretchAudioAuto classes that were previously used for EveStretch and EveStretch3 audio triggering. In addition this commit:

* Adds a new type of interfaced object, IStretchAudio, whose logic sits in carbon-audio and handles the actual implementation of the audio side of things.
* Correctly starts and stops the stretch audio. Previous logic for this was actually triggered in the wrong methods.

Originally, the intent of this work was to simply fix a bug where the stretch audio emitter was not following the camera (https://ccpgames.atlassian.net/browse/PLAT-10189), but it made more sense to move this logic to Carbon Audio now that we have the carbon-trinityaudioapi. Along with this it is more stable if the audio listener is used as a anchor point for the stretch audio emitter rather than the Trinity camera.